requirements.txt
fastapi
loguru
playwright
pydantic
python-stopwatch
python-tml